+++ src/rsfc/rsfc_core.py
@@ -1,4 +1,3 @@
-from somef import somef_cli
 from jinja2 import Environment, BaseLoader
 from datetime import datetime, timezone
 import os
@@ -6,35 +5,15 @@ import json
 from rsfc.model import assessedSoftware as soft
 from rsfc.rsfc_tests import rsfc_tests as rt
 from importlib.resources import files
-from rsfc.utils import constants, rsfc_helpers
-import io
-import contextlib
+from rsfc.harvesters import somef_harvester as som
 
 
-def somef_assessment(repo_url, threshold):
-    
-    print("Extracting repository metadata with SOMEF...")
-    
-    with contextlib.redirect_stdout(io.StringIO()), contextlib.redirect_stderr(io.StringIO()):
-        repo_data = somef_cli.cli_get_data(threshold=threshold, ignore_classifiers=True, repo_url=repo_url, readme_only=False)
-        
-    repo_data = json.loads(json.dumps(repo_data.results))
-    
-    #Guardar somef output
-    os.makedirs('./outputs', exist_ok=True)
-    with open('./outputs/somef_assessment.json', 'w', encoding='utf-8') as f:
-        json.dump(repo_data, f, indent=4, ensure_ascii=False)
-    
-    return repo_data
 
-
-def render_template(repo_url, repo_type, checks):
+def render_template(sw, checks):
     
     print("Rendering assessment...")
     
     data = dict()
-    
-    sw = soft.AssessedSoftware(repo_url, repo_type)
     
     data['name'] = sw.software_name
     data['url'] = sw.software_url
@@ -60,31 +39,25 @@ def render_template(repo_url, repo_type, checks):
 
 def build_assessment(repo_url):
     
-    somef_data = somef_assessment(repo_url, 0.8)
+    somef = som.SomefExtractor(repo_url)
+    somef_data = somef.somef_assessment(repo_url, 0.8)
+    sw = soft.AssessedSoftware(repo_url)
     
     print("Assessing repository...")
-    
-    if "github" in repo_url:
-        repo_type = constants.REPO_TYPES[0]
-        repo_branch = None
-    elif "gitlab" in repo_url:
-        repo_type = constants.REPO_TYPES[1]
-        repo_base_api_url = rsfc_helpers.get_repo_api_url(repo_url, repo_type)
-        repo_branch = rsfc_helpers.get_gitlab_default_branch(repo_base_api_url, repo_type)
 
     checks = [
         rt.test_has_license(somef_data),
         rt.test_license_spdx_compliant(somef_data),
-        rt.test_license_info_in_metadata_files(somef_data, repo_url, repo_type, repo_branch),
+        rt.test_license_info_in_metadata_files(somef_data, repo_url, sw.repo_type, sw.repo_branch),
         rt.test_authors_contribs(somef_data),
         rt.test_authors_orcids(somef_data),
-        rt.test_author_roles(repo_url, repo_type, repo_branch),
-        rt.test_metadata_exists(somef_data, repo_url, repo_type, repo_branch),
-        rt.test_codemeta_exists(repo_url, repo_type, repo_branch),
+        rt.test_author_roles(repo_url, sw.repo_type, sw.repo_branch),
+        rt.test_metadata_exists(somef_data, repo_url, sw.repo_type, sw.repo_branch),
+        rt.test_codemeta_exists(repo_url, sw.repo_type, sw.repo_branch),
         rt.test_descriptive_metadata(somef_data),
         rt.test_title_description(somef_data),
         rt.test_has_citation(somef_data),
-        rt.test_reference_publication(somef_data, repo_url, repo_type, repo_branch),
+        rt.test_reference_publication(somef_data, repo_url, sw.repo_type, sw.repo_branch),
         rt.test_software_documentation(somef_data),
         rt.test_readme_exists(somef_data),
         rt.test_contact_support_documentation(somef_data),
@@ -97,22 +70,22 @@ def build_assessment(repo_url):
         rt.test_latest_release_consistency(somef_data),
         rt.test_semantic_versioning_standard(somef_data),
         rt.test_version_scheme(somef_data),
-        rt.test_metadata_record_in_zenodo_or_software_heritage(somef_data, repo_url, repo_type, repo_branch),
+        rt.test_metadata_record_in_zenodo_or_software_heritage(somef_data, repo_url, sw.repo_type, sw.repo_branch),
         rt.test_id_presence_and_resolves(somef_data),
-        rt.test_id_associated_with_software(somef_data, repo_url, repo_type, repo_branch),
+        rt.test_id_associated_with_software(somef_data, repo_url, sw.repo_type, sw.repo_branch),
         rt.test_id_proper_schema(somef_data),
         rt.test_identifier_in_readme_citation(somef_data),
         rt.test_identifier_resolves_to_software(somef_data),
-        rt.test_presence_of_tests(repo_url, repo_type, repo_branch),
+        rt.test_presence_of_tests(repo_url, sw.repo_type, sw.repo_branch),
         rt.test_github_action_tests(somef_data),
         rt.test_repository_workflows(somef_data),
         rt.test_is_github_repository(repo_url),
-        rt.test_repo_enabled_and_commits(somef_data, repo_url, repo_type, repo_branch),
+        rt.test_repo_enabled_and_commits(somef_data, repo_url, sw.repo_type, sw.repo_branch),
         rt.test_has_tickets(somef_data),
         rt.test_repo_status(somef_data)
     ]
     
-    results = render_template(repo_url, repo_type, checks)
+    results = render_template(sw, checks)
     
     print("Saving assessment locally...")
